pyproject.toml
[project]
name = "webos"
version = "0.1.0"
description = "A Modular Monolith WebOS Framework"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "nicegui>=1.4.0",
    "beanie>=1.20.0",
    "motor>=3.3.0",
    "fastapi-users>=12.0.0",
    "fastapi-users-db-beanie>=5.0.0",
    "taskiq>=0.11.0",
    "taskiq-fastapi>=0.3.0",
    "taskiq-redis>=0.5.0",
    "pluggy>=1.0.0",
    "loguru>=0.7.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.24.0",
    "diskcache>=5.6.0",
    "slowapi>=0.1.8",
    "boto3>=1.28.0",
    "python-slugify>=8.0.0",
    "Pillow>=10.2.0",
    "mutagen>=1.47.0",
    "PyMuPDF>=1.24.2",
    "python-docx>=1.1.0",
    "python-magic>=0.4.27; sys_platform != 'win32'",
    "python-magic-bin>=0.4.14; sys_platform == 'win32'",
    "watchdog>=4.0.0",
    "ultralytics>=8.1.0",
    "onnxruntime>=1.17.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
    "bandit>=1.7.0",
    "playwright>=1.30.0",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
